Migrating the open command over and pouring some sugar on it.
